Migrate workflow step sub-tabs from query params to real routes with per-tab data loading
Actions:
+ add basePath prop to SubTabStrip for route-based tabs instead of ?subtab= query params
+ add JSDoc explaining basePath enables server-rendered route links with clean deep-linking
+ derive currentValue from last pathname segment when basePath is set
+ update hrefFor to build route paths when basePath is provided
+ add shrink-0 class to SubTabStrip nav to prevent flexbox crushing on tall pages
