+++ ui/packages/comhairle/src/lib/components/SubTabStrip.svelte
@@ -14,23 +14,43 @@
 		 *   (e.g. Recruit/invites) → `bg-secondary`.
 		 */
 		tone?: 'primary' | 'secondary';
+		/**
+		 * When set, each tab is a route link (`${basePath}/${value}`) and the active tab is
+		 * matched by the last path segment instead of the `?subtab=` query param. Use this when
+		 * the sub-tabs are real routes (e.g. a workflow step's Configure/Setup/…), so the strip
+		 * server-renders and deep-links cleanly. Omit for query-param sub-tabs.
+		 */
+		basePath?: string;
 	};
 
-	let { items, paramName = 'subtab', defaultValue, tone = 'secondary' }: Props = $props();
+	let {
+		items,
+		paramName = 'subtab',
+		defaultValue,
+		tone = 'secondary',
+		basePath
+	}: Props = $props();
 
 	let currentValue = $derived(
-		page.url.searchParams.get(paramName) ?? defaultValue ?? items[0]?.value
+		basePath
+			? (page.url.pathname.replace(/\/+$/, '').split('/').pop() ?? '')
+			: (page.url.searchParams.get(paramName) ?? defaultValue ?? items[0]?.value)
 	);
 
 	function hrefFor(value: string): string {
+		if (basePath) return `${basePath}/${value}`;
 		const params = new URLSearchParams(page.url.searchParams);
 		params.set(paramName, value);
 		return `${page.url.pathname}?${params.toString()}`;
 	}
 </script>
 
+<!-- `shrink-0`: when this strip is a direct flex child of the scrolling `main` column (route
+	 sub-tabs, e.g. a workflow step), a tall page below it would otherwise let flexbox crush the
+	 strip to its 1px border. Harmless where it sits inside a `shrink-0` header block (Events). -->
 <nav
-	class="border-border scrollbar-none flex w-full overflow-x-auto border-b {tone === 'primary'
+	class="border-border scrollbar-none flex w-full shrink-0 overflow-x-auto border-b {tone ===
+	'primary'
 		? 'bg-secondary'
 		: 'bg-accent'}"
 	aria-label="Sub sections"
