+++ ui/packages/comhairle/src/routes/(admin)/admin/conversations/[conversation_id]/design/step/[step_id]/configure/+page.svelte
@@ -1,0 +1,12 @@
+<script lang="ts">
+	import CommonStepConfig from '$lib/components/CommonStepConfig/CommonStepConfig.svelte';
+
+	let { data } = $props();
+
+	let conversation = $derived(data.conversation);
+	let step = $derived(data.step);
+</script>
+
+{#if step}
+	<CommonStepConfig conversation_id={conversation.id} {conversation} {step} inline />
+{/if}
