+++ ui/packages/comhairle/src/routes/(admin)/admin/conversations/[conversation_id]/design/step/[step_id]/insights/+page.svelte
@@ -1,0 +1,15 @@
+<script lang="ts">
+	import PolisInsights from '$lib/tools/polis/PolisInsights.svelte';
+
+	let { data } = $props();
+
+	let step = $derived(data.step);
+</script>
+
+{#if step}
+	<PolisInsights
+		workflowStepId={step.id}
+		reportData={data.reportData ?? null}
+		statementAux={data.statementAux ?? []}
+	/>
+{/if}
