+++ ui/packages/comhairle/src/routes/(admin)/admin/conversations/[conversation_id]/design/step/[step_id]/moderation/+page.svelte
@@ -1,0 +1,11 @@
+<script lang="ts">
+	import PolisModeration from '$lib/tools/polis/PolisModeration.svelte';
+
+	let { data } = $props();
+
+	let step = $derived(data.step);
+</script>
+
+{#if step}
+	<PolisModeration workflowStepId={step.id} statements={data.statementAux ?? []} />
+{/if}
